refactor(block-cow): move get_size_sectors to blockdev module
The function wraps `blockdev --getsz`, not `losetup`. Give it its own
module to match the losetup/dmsetup naming convention.
